requirements.txt
Flask==3.0.2
groq==0.4.0
python-docx==1.1.0
twilio==9.0.2
pillow==10.2.0
